don't use port forwarding with Virtualbox as I discovered the hard way, Virtualbox port forwarding + Docker port forwarding interacts badly and randomly messes up HTTP requests (yes, really).
this manifests with a bunch of err::CONTENT_TYPE_LENGTH_MISMATCH errors.

this is fixed by simply not doing port forwarding and only doing host-only networking.
this is one shade less convenient for dev deployments as the user needs to learn the IP address of the container
this does inadvertently allow for one cool thing (Bob will be stoked): you can now access the VMs for testing from another computer

more details:
https://github.com/boot2docker/boot2docker-cli/issues/164
https://groups.google.com/forum/#!topic/docker-user/bvZdFqQnUK8 <-- we were experiencing this content-length error